+++ ACKategoriesCore/StringExtensions.swift
@@ -18,9 +18,14 @@ extension String {
         return !isEmpty && rangeOfCharacter(from: CharacterSet.decimalDigits.inverted) == nil
     }
 
-    /// Uses self as key to Localizable.strings and returns its localized value or self
-    public func localized() -> String {
-        return NSLocalizedString(self, comment: "")
+    /// Uses `self` as key in stringsfile and returns its localized value or `self`
+    ///
+    /// - parameter comment: Comment to string usage, is displayed if `key` is missing in table and `value` is empty string
+    /// - parameter value: Value to be displayed if `key` is missing, if empty string then `key` is displayed
+    /// - parameter tableName: Name of strings table
+    /// - parameter bundle: Bundle where localization table is stored
+    public func localized(comment: String = "", value: String = "", tableName: String? = nil, bundle: Bundle = .main) -> String {
+        NSLocalizedString(self, tableName: tableName, bundle: bundle, value: value, comment: comment)
     }
 
     /// Normalizes string - removes interpuction etc.